main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import Optional, List
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(
    title="GORA HRMS AI Backend",
    description="AI-powered features for HRMS",
    version="1.0.0"
)

# CORS configuration
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Update with specific origins in production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Models
class HealthResponse(BaseModel):
    status: str
    message: str

class AIQuery(BaseModel):
    query: str
    context: Optional[dict] = None

class AIResponse(BaseModel):
    success: bool
    data: Optional[dict] = None
    message: str

# Routes
@app.get("/", response_model=HealthResponse)
async def root():
    return {
        "status": "ok",
        "message": "GORA HRMS AI Backend is running"
    }

@app.get("/health", response_model=HealthResponse)
async def health_check():
    return {
        "status": "ok",
        "message": "Service is healthy"
    }

# Future AI endpoints
@app.post("/api/ai/analyze-attendance", response_model=AIResponse)
async def analyze_attendance(query: AIQuery):
    """
    Analyze attendance patterns using AI
    Future implementation: Use ML models to detect patterns, anomalies
    """
    return {
        "success": True,
        "data": {
            "analysis": "AI analysis will be implemented here",
            "insights": []
        },
        "message": "Analysis endpoint ready for implementation"
    }

@app.post("/api/ai/predict-leaves", response_model=AIResponse)
async def predict_leaves(query: AIQuery):
    """
    Predict leave patterns and trends
    Future implementation: Use historical data to predict leave requests
    """
    return {
        "success": True,
        "data": {
            "predictions": [],
            "trends": []
        },
        "message": "Prediction endpoint ready for implementation"
    }

@app.post("/api/ai/recommend-actions", response_model=AIResponse)
async def recommend_actions(query: AIQuery):
    """
    Recommend HR actions based on data analysis
    Future implementation: Use AI to suggest HR interventions
    """
    return {
        "success": True,
        "data": {
            "recommendations": []
        },
        "message": "Recommendation endpoint ready for implementation"
    }

@app.post("/api/ai/chat", response_model=AIResponse)
async def ai_chat(query: AIQuery):
    """
    AI-powered chat assistant for HR queries
    Future implementation: Integrate with LLM for natural language queries
    """
    return {
        "success": True,
        "data": {
            "response": "AI chat will be implemented here",
            "suggestions": []
        },
        "message": "Chat endpoint ready for implementation"
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
