requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9

# Future AI dependencies (uncomment when needed)
# openai==1.51.2
# langchain==0.3.7
# pgvector==0.3.5
# numpy==2.1.3
# pandas==2.2.3
# scikit-learn==1.5.2
